Restructuring of structures.graphs.
